fix compilation errors on gcc 8.3.1

+++ src/lib/x-ray/includes/device.hh
@@ -190,8 +190,8 @@ namespace penred{
 
       bool printGeo;
 
-      unsigned detBinsX, detBinsY;
-      unsigned eBins;
+      unsigned long detBinsX, detBinsY;
+      unsigned long eBins;
 
       double tolerance;
